requirements.txt
faster-whisper==1.0.3
pyttsx3==2.90
pyyaml==6.0.2
typer==0.12.5
rich==13.9.2
pydub==0.25.1
requests==2.31.0
